// src/app/gateway/models/Module.Dto.ts
// @ts-nocheck


export class ModuleDto {
    delayRetryTime?: number;
    description?: string;
    esbMode?: number;
    limitForPeriod?: number;
    limitRefreshPeriod?: number;
    moduleAuthMode?: number;
    moduleBase?: boolean;
    moduleGroup?: number;
    moduleId?: number;
    moduleTitle?: string;
    moduleType?: number;
    partyId?: number;
    partyTitle?: string;
    retryCount?: number;
    status?: number;
    partyBase?: boolean;
}